+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # DMP Tool Apollo Server Change Log
 
 ### Added
+- Added tables to support template customizations
 - Added relatedWorks, workVersions, and works tables to database and two stored procedures that will be used to update the related works.
 - Added tmp dir for jest caching (ignored by git)
 - Added Trivy scan and `npm run build` to precommit hooks